netlify/functions/requirements.txt
google-generativeai==0.3.2
pypdf2==3.0.1
python-dotenv==0.19.2
elevenlabs==0.2.27
